+++ spec/wavefront-sdk/metric_helper_spec.rb
@@ -1,8 +1,6 @@
 #!/usr/bin/env ruby
 
 require 'minitest/autorun'
-require 'spy'
-require 'spy/integration'
 require_relative '../spec_helper'
 require_relative '../../lib/wavefront-sdk/metric_helper'
 
@@ -136,7 +134,7 @@ class WavefrontMetricHelperTest < MiniTest::Test
 
     out = wf.flush_gauges(input)
     args = spy.calls.first.args.first
-    assert_instance_of(Mocket, out)
+    assert_instance_of(Wavefront::Response, out)
     assert spy.has_been_called?
     assert_equal(input, args)
     assert(args.any? { |a| a.key?(:tags) && a[:tags] == WH_TAGS })
@@ -150,7 +148,7 @@ class WavefrontMetricHelperTest < MiniTest::Test
     mocket = BadMocket.new
     spy = Spy.on(wf.writer.writer, :write).and_return(mocket)
     out = wf.flush_gauges(input)
-    assert_instance_of(BadMocket, out)
+    assert_instance_of(Wavefront::Response, out)
     wf.gauge('m2.p', 9)
     wf.gauge('m3.p', 9)
     assert spy.has_been_called?
@@ -173,7 +171,7 @@ class WavefrontMetricHelperTest < MiniTest::Test
     out = wf.flush_counters(input)
     args = spy.calls.first.args.first
 
-    assert_instance_of(Mocket, out)
+    assert_instance_of(Wavefront::Response, out)
     assert spy.has_been_called?
     assert_equal(3, args.size)
 
@@ -201,7 +199,7 @@ class WavefrontMetricHelperTest < MiniTest::Test
     out = wf.flush_counters(input)
     args = spy.calls.first.args.first
 
-    assert_instance_of(BadMocket, out)
+    assert_instance_of(Wavefront::Response, out)
     assert spy.has_been_called?
     assert_equal(3, args.size)
 
@@ -232,7 +230,7 @@ class WavefrontMetricHelperTest < MiniTest::Test
     out = wfd.flush_dists(input)
     args = spy.calls.first.args.first
 
-    assert_instance_of(Mocket, out)
+    assert_instance_of(Wavefront::Response, out)
     assert spy.has_been_called?
     assert_equal(3, args.size)
 
@@ -261,7 +259,7 @@ class WavefrontMetricHelperTest < MiniTest::Test
     out = wfd.flush_dists(input)
     args = spy.calls.first.args.first
 
-    assert_instance_of(BadMocket, out)
+    assert_instance_of(Wavefront::Response, out)
     assert spy.has_been_called?
     assert_equal(3, args.size)
     refute_empty(wfd.buf[:dists])